requirements.txt
boto3==1.14.32
botocore==1.17.32
certifi==2020.6.20
chardet==3.0.4
dj-database-url==0.5.0
Django==2.1.7
django-appconf==1.0.3
django-cleanup==3.2.0
django-extensions==3.0.3
django-filebrowser-no-grappelli==3.7.7
django-imagekit==4.0
django-jet==1.0.8
django-postgres-copy==2.4.2
django-smart-selects==1.5.6
django-storages==1.6.5
django-tinymce4-lite==1.8.0
django-widget-tweaks==1.4.8
docutils==0.14
filebrowser-safe==0.5.0
future==0.18.2
gunicorn==19.9.0
idna==2.10
jmespath==0.9.4
jsmin==2.2.2
pilkit==2.0
Pillow==5.4.1
psycopg2==2.8.3
psycopg2-binary==2.8.2
#Pillow
#psycopg2
#psycopg2-binary
PyJWT==1.7.1
python-dateutil==2.8.0
python-decouple==3.3
pytz==2018.9
PyVimeo==0.3.10
requests==2.24.0
s3transfer==0.3.3
six==1.12.0
tinydb==4.1.1
tuspy==0.2.4
twilio==6.44.1
urllib3==1.25.6
whitenoise==4.1.2
xxhash==0.4.3